Fix all 82 ruff linting errors
Fixes include:
- PIE794: Fix duplicate class field (RX -> RZ typo)
- SIM105: Replace try/except/pass with contextlib.suppress
- RUF059: Prefix unused unpacked variables with underscore
- PERF102: Use .values() instead of .items() when only values needed
- SIM103: Return condition directly instead of if/else True/False
- C419: Remove unnecessary list comprehension in all()
- C416: Replace identity list comprehension with list()
- RUF017: Replace quadratic sum() with list comprehension
- RUF003: Add noqa for intentional unicode characters
- RUF013: Fix implicit Optional type annotations
- SIM401: Replace if/else with dict.get()
- FURB110: Replace ternary with or operator
- SIM201/SIM202: Use != instead of not ==, == instead of not !=
- RUF046: Remove unnecessary int() around round()
- F811: Add noqa for intentional class field redefinition
- PERF401: Convert manual list append loops to comprehensions
- RUF005: Use iterable unpacking instead of concatenation
- SIM118: Remove unnecessary .keys() calls
- SIM115: Add noqa for valid NamedTemporaryFile usage
- E501: Break long lines
- C417: Replace map(lambda) with generator expressions
- SIM223: Simplify always-False condition
- E721: Use 'is' for type comparisons

+++ lib/logitech_receiver/onboard_eq.py
@@ -66,7 +66,7 @@ def _quantize_coeffs(b0, b1, b2, a1, a2):
     scales = [2**31, 2**30, 2**31, 2**30, 2**31]  # b0, b1, b2, a1, a2
     words = []
     for val, scale in zip([b0, b1, b2, a1, a2], scales):
-        q = int(round(val * scale))
+        q = round(val * scale)
         q = max(-(1 << 31), min((1 << 31) - 1, q))
         q = q & 0xFFFFFF00  # 24-bit precision (low byte always zero)
         words.append((q >> 16) & 0xFFFF)  # high word
@@ -103,7 +103,7 @@ def _build_coeff_section(bands, sample_rate, section_type=1):
         all_words.extend(_quantize_coeffs(b0 / rescale, b1 / rescale, b2 / rescale, a1, a2))
 
     # Rescale factor as Q6.26, 24-bit precision
-    rs = int(round(rescale * (1 << 26)))
+    rs = round(rescale * (1 << 26))
     rs = max(-(1 << 31), min((1 << 31) - 1, rs)) & 0xFFFFFF00
     all_words.append((rs >> 16) & 0xFFFF)
     all_words.append(rs & 0xFFFF)
